Refactored service profiles. Having a Service Profile just for one registration doesn't make sense. Added Service Profile in the DigitalOcean provider. Makes sense to keep all the registrations in that profile. That way, all the DigitalOcean-related code is in the DigitalOcean folder.
