+++ src/DDNSUpdate/Application/Providers/DigitalOcean/DigitalOceanServiceProfile.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using ServiceProfiles;
+
+namespace DDNSUpdate.Application.Providers.DigitalOcean
+{
+    public class DigitalOceanServiceProfile : HostServiceProfile
+    {
+        public override void Configure(IHostServiceProfileContext context)
+        {
+            context.Services.AddScoped<IDDNSService, DigitalOceanDDNSService>();
+        }
+    }
+}
