--- src/DDNSUpdate/Infrastructure/ServiceProfiles/ExternalAddressProfile.cs
@@ -1,14 +1,0 @@
-﻿using ServiceProfiles;
-using Microsoft.Extensions.DependencyInjection;
-using DDNSUpdate.Application.ExternalAddresses;
-
-namespace DDNSUpdate.Infrastructure.ServiceProfiles
-{
-    public class ExternalAddressProfile : HostServiceProfile
-    {
-        public override void Configure(IHostServiceProfileContext context)
-        {
-            context.Services.AddHttpClient<IExternalAddressClient, ExternalAddressClient>();
-        }
-    }
-}
