--- src/DDNSUpdate/Infrastructure/ServiceProfiles/ValidationServiceProfile.cs
@@ -1,16 +1,0 @@
-﻿using ServiceProfiles;
-using FluentValidation;
-using DDNSUpdate.Application.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace DDNSUpdate.Infrastructure.ServiceProfiles
-{
-    public class ValidationServiceProfile : HostedServiceProfile
-    {
-        public override void Configure(IHostServiceProfileContext context)
-        {
-            context.Services.AddValidatorsFromAssembly(ThisAssembly);
-            context.Services.AddScoped<IConfigurationValidator, ConfigurationValidator>();
-        }
-    }
-}
